requirements.txt
fastapi
uvicorn
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-openai>=0.1.0
openai
faiss-cpu
sentence-transformers
python-multipart
streamlit
python-dotenv
requests
pypdf
pdfplumber
python-docx
ragas
